feat: 1. add convert-chinese....py script 2. add virtual active alias 3. change font of raw term 4. make populate file check file existence

+++ dot-home/.local/populate-dotfiles.sh
@@ -8,12 +8,13 @@ src_dir=$(realpath $(pwd)/$script_dir/bin)
 
 mkdir -p $dest_dir
 
-ln -sv $src_dir/auto_rip.sh $dest_dir/auto-rip
-ln -sv $src_dir/container_as_cmd.sh $dest_dir/container_as_cmd 
-ln -sv $src_dir/hdmi-hotplug.sh $dest_dir/
-ln -sv $src_dir/i3-move-position.sh $dest_dir/
-ln -sv $src_dir/i3-utils-cmd-term.sh $dest_dir/
-ln -sv $src_dir/net-wrapper.sh $dest_dir/netrp
-ln -sv $src_dir/tmuxpcmd.sh $dest_dir/tcmd
-ln -sv $src_dir/.nft/bin/v2ray-wrapper.nft.sh $dest_dir/v2ray-nft 
-ln -sv $src_dir/vm-win-11 $dest_dir/vm-w11
+[[ ! -L  "$dest_dir/auto-rip" ]] && ln -sv $src_dir/auto_rip.sh $dest_dir/auto-rip
+[[ ! -L  "$dest_dir/container_as_cmd" ]] && ln -sv $src_dir/container_as_cmd.sh $dest_dir/container_as_cmd 
+[[ ! -L  "$dest_dir/hdmi-hotplug.sh" ]] && ln -sv $src_dir/hdmi-hotplug.sh $dest_dir/
+[[ ! -L  "$dest_dir/i3-move-position.sh" ]] && ln -sv $src_dir/i3-move-position.sh $dest_dir/
+[[ ! -L  "$dest_dir/i3-utils-cmd-term.sh" ]] && ln -sv $src_dir/i3-utils-cmd-term.sh $dest_dir/
+[[ ! -L  "$dest_dir/netrp" ]] && ln -sv $src_dir/net-wrapper.sh $dest_dir/netrp
+[[ ! -L  "$dest_dir/tcmd" ]] && ln -sv $src_dir/tmuxpcmd.sh $dest_dir/tcmd
+[[ ! -L  "$dest_dir/v2ray-nft" ]] && ln -sv $src_dir/.nft/bin/v2ray-wrapper.nft.sh $dest_dir/v2ray-nft 
+[[ ! -L  "$dest_dir/vm-w11" ]] && ln -sv $src_dir/vm-win-11 $dest_dir/vm-w11
+[[ ! -L  "$dest_dir/convert-chinese-2pinyin-dict.py" ]] && ln -sv $src_dir/convert-chinese-2pinyin-dict.py $dest_dir/
